Changes to support journaling.
git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk@5806 1a4d904c-e006-0410-be0b-b506c45ff692

+++ simias/src/Sync/SyncService.cs
@@ -726,12 +726,16 @@ namespace Simias.Sync
 						}
 						else
 						{
-							// If this is a file delete the file.
-							BaseFileNode bfn = node as BaseFileNode;
-							if (bfn != null)
+							// Don't remove store managed files (this is handled in the commit code).
+							if ( !node.IsType( NodeTypes.StoreFileNodeType ) )
 							{
-								name = bfn.GetFullPath(collection);
-								SyncFile.DeleteFile(collection, bfn, name);
+								// If this is a file delete the file.
+								BaseFileNode bfn = node as BaseFileNode;
+								if (bfn != null)
+								{
+									name = bfn.GetFullPath(collection);
+									SyncFile.DeleteFile(collection, bfn, name);
+								}
 							}
 							collection.Delete(node);
 							collection.Commit(node);
